runner/minitest.rb requires correct runner
