+++ lib/test_queue/runner/minitest5.rb
@@ -45,20 +45,6 @@ module TestQueue
         ::MiniTest::Test.runnables = iterator
         ::MiniTest.run ? 0 : 1
       end
-
-      def summarize_worker(worker)
-        worker.stats.each do |s, val|
-          stats[s.to_s] = val
-        end
-
-        num_tests = worker.lines.grep(/, \d+ errors?, /).first
-        failures  = worker.lines.select{ |line|
-          line if (line =~ /^Finished/) ... (line =~ /, \d+ errors?, /)
-        }[1..-2]
-        failures = failures.join("\n") if failures
-
-        [ num_tests, failures ]
-      end
     end
   end
 end
